refactor: 공유 바텀시트 Footer 컴포넌트명 변경
